moved config to docs/index.js
