+++ server/src/server.js
@@ -1,25 +1,10 @@
 const http = require('http');
 const express = require('express');
 const swaggerUI = require('swagger-ui-express');
-const swaggerJsDoc = require('swagger-jsdoc');
 const config = require('./config');
 const { apiRouter, mainRouter, authRouter } = require('./routers');
 const setupMiddlewares = require('./middlewares');
-
-const swaggerDefinition = {
-  openapi: '3.0.0',
-  info: {
-    title: 'Modern-and-Magic-Chess-Online',
-    version: '1.0.0',
-  },
-};
-
-const options = {
-  swaggerDefinition,
-  apis: ['./src/docs/*.js'],
-};
-
-const swaggerSpec = swaggerJsDoc(options);
+const swaggerSpec = require('./docs/index');
 
 const app = express();
